refactor(welcome): split splash into colocated splash/ components
Break splash.vue into splash/index.vue (orchestrator) plus splash-nav,
splash-copy, and splash-preview, each owning its own state.
