+++ src/views/welcome/index.vue
@@ -4,7 +4,7 @@ import { useSessionStore } from '@/stores/session'
 import { onMounted, useTemplateRef } from 'vue'
 import { emitSfx } from '@/sfx/bus'
 import { useModal } from '@/composables/modal'
-import Splash from './splash.vue'
+import Splash from './splash/index.vue'
 import SectionFeatures from './section-features.vue'
 import SectionConfig from './section-config.vue'
 import SectionPricing from './section-pricing.vue'
